Traktor: Heightfield attributes as individual layers including density per attribute. Much finer control of grass, undergrowth and rubble.

+++ code/Terrain/Editor/AttributeBrush.h
@@ -39,7 +39,7 @@ private:
 	int32_t m_radius;
 	const IFallOff* m_fallOff;
 	float m_strength;
-	int32_t m_attribute;
+	uint8_t m_attribute;
 };
 
 }
